Stock: updateLocation + jav + refakt

+++ bbxBE.Infrastructure.Persistence/Migrations/0042_CrFld_Location.cs
@@ -19,7 +19,7 @@ namespace bbxBE.Infrastructure.Persistence.Migrations
         {
 
             Alter.Table("Stock")
-                   .AddColumn("LocationID").AsDecimal().Nullable();
+                   .AddColumn("LocationID").AsInt64().Nullable();
         }
     }
 }
